Add outlier removal in initialise_vo
Add a new file utils.py containing functions for outlier removal.
Applied in initialise_vo.
Make the aspect ration of the landmark visualisation adaptable, the
default is now closer to the proportions in the image.
Create a method draw_all in Bootstrap that shows both the keypoints and
the landmarks.